Fixed embed deletion

+++ inc/admin.posting.php
@@ -24,7 +24,7 @@ function deletePostMod($conn, $board, $postno, $onlyimgdel = 0)
 					{
 						$filename = substr($filename, 8);
 					}
-					if (substr($filename, 0, 6) != "embed:")
+					if ((substr($filename, 0, 6) != "embed:") && ($filename != "deleted"))
 					{
 						unlink("./".$board."/src/".$filename);
 						unlink("./".$board."/src/thumb/".$filename);
@@ -53,7 +53,7 @@ function deletePostMod($conn, $board, $postno, $onlyimgdel = 0)
 						{
 							$filename = substr($filename, 8);
 						}
-						if (substr($filename, 0, 6) != "embed:")
+						if ((substr($filename, 0, 6) != "embed:") && ($filename != "deleted"))
 						{
 							unlink("./".$board."/src/".$filename);
 							unlink("./".$board."/src/thumb/".$filename);
@@ -66,7 +66,7 @@ function deletePostMod($conn, $board, $postno, $onlyimgdel = 0)
 						{
 							$filename = substr($filename, 8);
 						}
-						if (substr($filename, 0, 6) != "embed:")
+						if ((substr($filename, 0, 6) != "embed:") && ($filename != "deleted"))
 						{
 							unlink("./".$board."/src/".$filename);
 							unlink("./".$board."/src/thumb/".$filename);
@@ -87,7 +87,7 @@ function deletePostMod($conn, $board, $postno, $onlyimgdel = 0)
 						{
 							$filename = substr($filename, 8);
 						}
-						if (substr($filename, 0, 6) != "embed:")
+						if ((substr($filename, 0, 6) != "embed:") && ($filename != "deleted"))
 						{
 							unlink("./".$board."/src/".$filename);
 							unlink("./".$board."/src/thumb/".$filename);
